// src/app/module/auth/auth.interface.ts
import { UserRole } from '../../lib/prisma';

export interface ILoginUserPayload {
  email: string;
  password: string;
}

export interface IRegisterUserPayload {
    name: string
    email: string
    password: string
    phone?: string;
    role?: UserRole;
}

export interface IRequestUser {
    userId: string
    email: string
    name: string
    role: UserRole
}